feat(jdsl.groovy): add support for configuring job by URL instead of scanning credentials
The change was made to the `jdsl.groovy` file. The `createJob` function now includes a new line `configuredByUrl(true)`. This change allows the job to be configured using a URL instead of scanning credentials. The purpose of this change is to simplify the job configuration process and improve security by avoiding the need to store and manage credentials.