feat(jdsl.groovy): add support for remote Jenkins file configuration
The changes in this commit add support for configuring a remote Jenkins file in the JDSL script. The remote Jenkins file "pipelines/nodejs.groovy" is now specified in the script. Additionally, a Git SCM configuration is added to fetch the Jenkins file from the specified GitHub repository "https://github.com/felix-orduz/jenkins.git". This allows for greater flexibility in managing and versioning Jenkins files for different pipelines.