feat(JCASC): add Jenkins Configuration as Code (JCASC) files and Docker configuration
- Add `casc.yaml` file to configure Jenkins settings such as the Jenkins URL, system message, security realm, and script approvals.
- Add `initialConfig.groovy` file to set the number of executors and mark the initial setup as completed.
- Add `plugins.txt` file to specify the required Jenkins plugins.
- Add `seed.jenkinsfile` to define a pipeline job for the Seed Job.
- Add `docker-compose.yml` file to configure the Jenkins and SSH agent services in Docker.
- Add `jenkins_master.Dockerfile` to build the Jenkins master Docker image, install Docker CLI, and copy the JCASC files.

These changes are made to automate the Jenkins setup and configuration process using JCASC and Docker, providing a reproducible and scalable environment for Jenkins.

chore(plugins_no_upload.txt): remove unused plugins from the project

The plugins_no_upload.txt file was deleted as it contained a list of unused plugins that were no longer needed in the project. This commit removes the file and cleans up the project by removing the unnecessary plugins.